Move test into config_test
